// src/api/generated/models/visitResponseBFF.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * BFF Service
 * Backend For Frontend service aggregating microservices
 * OpenAPI spec version: 1.0.0
 */
import type { VisitResponseBFFNotasVisita } from "./visitResponseBFFNotasVisita";
import type { VisitResponseBFFRecomendaciones } from "./visitResponseBFFRecomendaciones";
import type { VisitResponseBFFArchivosEvidencia } from "./visitResponseBFFArchivosEvidencia";

/**
 * Response schema for a single visit (BFF).
 */
export interface VisitResponseBFF {
  id: string;
  seller_id: string;
  client_id: string;
  fecha_visita: string;
  status: string;
  notas_visita: VisitResponseBFFNotasVisita;
  recomendaciones: VisitResponseBFFRecomendaciones;
  archivos_evidencia: VisitResponseBFFArchivosEvidencia;
  client_nombre_institucion: string;
  client_direccion: string;
  client_ciudad: string;
  client_pais: string;
  created_at: string;
  updated_at: string;
}